+++ src/lang/lang.mli
@@ -57,7 +57,6 @@ and in_value = Value.in_value =
   | Tuple of value list
   | Null
   | Meth of string * value * value
-  | Ref of value Atomic.t
   | Fun of (string * string * value option) list * lazy_env * Term.t
   (* A function with given arguments (argument label, argument variable, default
      value), closure and value. *)
@@ -128,7 +127,11 @@ val to_valued_option : (value -> 'a) -> value -> 'a option
 val to_default_option : default:'a -> (value -> 'a) -> value -> 'a
 val to_product : value -> value * value
 val to_tuple : value -> value list
-val to_ref : value -> value Atomic.t
+val to_ref : value -> (unit -> value) * (value -> unit)
+
+val to_valued_ref :
+  (value -> 'a) -> ('a -> value) -> value -> (unit -> 'a) * ('a -> unit)
+
 val to_string_list : value -> string list
 val to_int_list : value -> int list
 val to_fun : value -> (string * value) list -> value
@@ -183,7 +186,7 @@ val product : value -> value -> value
 val tuple : value list -> value
 val meth : value -> (string * value) list -> value
 val record : (string * value) list -> value
-val reference : value Atomic.t -> value
+val reference : (unit -> value) -> (value -> unit) -> value
 
 (** Build a function from an OCaml function. Items in the prototype indicate
     the label and optional values. *)